Pull the :id column out when necessary.
Some databases -- and postgres is one of these, firebird is another --
implement .getGeneratedKeys() to simply return the record.
Typically, that is because the db in question is unable to easily
determine which columns are autogenerated ones, because reasons.

So when a whole entire record map comes back, we pull :id off of it,
because we know (since we created it) which column we're interested in.
Kind of just one of those things.

postgres mailing list discussion:
  http://www.postgresql.org/message-id/004801cb7518$cbc632e0$635298a0$@pravdin@disi.unitn.it